pyproject.toml
[tool.poetry]
name = "smartchatai-backend"
version = "1.0.0"
description = "SmartChatAI Backend - FastAPI backend with AI chat, document analysis, and resume tools"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
firebase-admin = "^6.2.0"
google-cloud-firestore = "^2.13.0"
google-auth = "^2.25.0"
requests = "^2.31.0"
httpx = "^0.25.0"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
langchain-huggingface = "^0.0.1"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
pypdf = "^3.17.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

